+++ changelog.md
@@ -1,5 +1,12 @@
 # eslint-config-env changelog
 
+## Next
+
+### Major
+
+- Updated the [`eslint`](https://npm.im/eslint) peer dependency to `^7.5.0`, due to the requirements of [`@babel/eslint-parser`](https://npm.im/@babel/eslint-parser).
+- Replaced [`babel-eslint`](https://npm.im/babel-eslint) with [`@babel/eslint-parser`](https://npm.im/@babel/eslint-parser).
+
 ## 17.0.0
 
 ### Major